+++ common/namespace/nsmanager/attr_validator_test.go
@@ -22,7 +22,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package nsattr
+package nsmanager
 
 import (
 	"testing"
