+++ common/namespace/nsmanager/errors.go
@@ -22,7 +22,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package nsattr
+package nsmanager
 
 import (
 	"go.temporal.io/api/serviceerror"
